+++ doc/release_notes.rst
@@ -25,6 +25,29 @@ Upcoming Release
     )
     plt.colorbar(plt.cm.ScalarMappable(cmap='viridis', norm=norm))
 
+* New utility functions to add legends for line widths (:func:`pypsa.plot.add_legend_lines`), circles and pie chart areas (:func:`pypsa.plot.add_legend_circles`), and patch colors (:func:`pypsa.plot.add_legend_patches`).
+  See the following example:
+
+  .. code-block:: python
+    :caption: Legend plotting example
+
+    import pypsa
+    import matplotlib.pyplot as plt
+    import cartopy.crs as ccrs
+    from pypsa.plot import add_legend_circles
+
+    n = pypsa.examples.ac_dc_meshed()
+
+    fig, ax = plt.subplots(subplot_kw={"projection": ccrs.PlateCarree()})
+    n.plot(ax=ax, bus_sizes=1)
+
+    add_legend_circles(
+        ax,
+        [1, 0.5],
+        ["reference size", "reference size 2"],
+        legend_kw=dict(frameon=False, bbox_to_anchor=(1,0.1))
+    )
+
 
 PyPSA 0.19.3 (22nd April 2022)
 ==============================