+++ doc/release_notes.rst
@@ -7,6 +7,8 @@ Upcoming Release
 
 .. warning:: The features listed below are not released yet, but will be part of the next release! To use the features already you have to install the ``master`` branch, e.g. ``pip install git+https://github.com/pypsa/pypsa#egg=pypsa``.
 
+* When iterating over components of a Subnetwork, only a those assets are included in the dataframes which are included in the subnetwork.
+
 * In ``n.plot()``, compute boundaries in all cases for consistent circle sizes. This is realised by setting a new default margin of 0.05.
 
 * Improved support for individually normed colorbars in ``n.plot()`` for buses, lines, links, transformers with keyword arguments ``bus_norm``, ``line_norm``, ``link_norm``, ``transformer_norm``.
@@ -48,6 +50,7 @@ Upcoming Release
         legend_kw=dict(frameon=False, bbox_to_anchor=(1,0.1))
     )
 
+* Compatibility with pyomo 6.4.1.
 
 PyPSA 0.19.3 (22nd April 2022)
 ==============================